LayoutTests/webgl/offscreen-webgl-errors-to-console-expected.txt

CONSOLE MESSAGE: WebGL: INVALID_ENUM: activeTexture: texture unit out of range
PASS Case: true
PASS Case: false
PASS successfullyParsed is true

TEST COMPLETE
